AF-112 - middleware-manager <- updated tests to match current state
